a2a-directory-agent/pyproject.toml
[project]
name = "a2a-directory-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn",
    "google-adk",
    "a2a-sdk",
    "python-dotenv",
    "toolbox-core",
]
